Add ThreadPool to limit threads usage
With this commit I added a new thread pool called `Rocketman::Pool`.
Currently it's hardcoded to spawn 5 workers, and each worker will pop
off the queue every 3 seconds. I plan to refactor these to be
customizable.

Due to this change in `Thread` strategy, it also required some
refactoring on the testing side of things. Since it's difficult to test
threaded code, I've decided to swap out the implementation of
`Rocketman::Pool` to `Rocketman::TestPool`, which defines the same
interface, except it is not a singleton and does not run asynchronously.