Add Redis as producer
With this commit, now clients are able to also listen to events emitted
from Redis (with Pub/Sub)!

This works because Rocketman now exposes a Bridge, and Bridge is responsible
for listening to events from the outside world (e.g: Redis) on behalf of
you. After listening to those events, Bridge will push those events onto
the internal Registry, which will then get queued and ready to be picked
up by other consumers.

This pattern is powerful because your consumers don't know where the
events are coming from, this gives you great advantage for when you
decide to move your message queue off of Ruby.