Add Redis as a persistence layer
By default, the events emitted from your app will be stored in an
in-memory `Queue`, which will get processed by Rocketman threaded
workers.

However this also means that if your app dies with events still in your
job queue, your emitted events which are stored in-memory will be lost.

That is obviously not desirable, so that's why Rocketman (this commit) ships with an
option to use `Redis` as your backing storage mechanism.